Building out the Distance / Bearing object
 * extended DistanceBearing to produce distance in miles (WiP)
 * improving data model
 * refactored tests to evaluate both object and Json encoding

 Changes to be committed:
	modified:   pom.xml
	modified:   src/main/java/com/daviddemartini/stratux/miltracker/comms/stratux/TrafficWebSocket.java
	modified:   src/main/java/com/daviddemartini/stratux/miltracker/datamodel/AircraftSBS.java
	new file:   src/main/java/com/daviddemartini/stratux/miltracker/geolocation/DistanceBearing.java
	modified:   src/test/java/com/daviddemartini/stratux/miltracker/AircraftSBSTest.java
	new file:   src/test/java/com/daviddemartini/stratux/miltracker/DistanceBearingTest.java

+++ src/main/java/com/daviddemartini/stratux/miltracker/datamodel/AircraftSBS.java
@@ -49,8 +49,8 @@ public class AircraftSBS {
     private int altitude;
     private float speedGround;
     private float track;
-    private float latitude;
-    private float longitude;
+    private double latitude;
+    private double longitude;
     private int verticalRate;
     private int squawkInt;
     private boolean squawkChanged;
@@ -134,10 +134,10 @@ public class AircraftSBS {
                         this.track = Float.parseFloat(fieldVal);
                         break;
                     case 14:
-                        this.latitude = Float.parseFloat(fieldVal);
+                        this.latitude = Double.parseDouble(fieldVal);
                         break;
                     case 15:
-                        this.longitude = Float.parseFloat(fieldVal);
+                        this.longitude = Double.parseDouble(fieldVal);
                         break;
                     case 16:
                         this.verticalRate = Integer.parseInt(fieldVal);
@@ -239,11 +239,11 @@ public class AircraftSBS {
         return track;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
